% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-nearest-neighbors cross-validation function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Data frame input of training data.}

\item{cl}{Vector input of true class values for the training data.}

\item{k_nn}{Integer representing the number of neighbors.}

\item{k_cv}{Integer representing the number of folds.}
}
\value{
List including a vector of class predictions for all observations and
a numeric with the cross-validation misclassification error.
}
\description{
Predicts the class of the input data using k-nearest-neighbors and performs
cross-validation to determine the misclassification error of the model.
}
\examples{
penguins <- my_penguins \%>\% tidyr::drop_na()
train <- penguins[,3:6]
cl <- penguins[,1]
my_knn_cv(train, cl, 1, 5)

}
\keyword{inference}
\keyword{prediction}
