% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear model function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Formula class object specifying relationship.}

\item{data}{Data frame input.}
}
\value{
Data frame including the estimate, standard error, t-value, and
p-value for each coefficient.
}
\description{
Fits a linear model to input data and outputs a table describing the estimated
coefficients.
}
\examples{
data(mtcars)
my_lm(mpg ~ ., data = mtcars)

}
\keyword{inference}
