% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random forest cross-validation function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Integer representing the number of folds.}
}
\value{
A numeric with the cross-validation error.
}
\description{
Using the palmerpenguins data, predicts body mass with covariates bill length,
bill depth, and flipper length by the random forest algorithm and performs
cross-validation on the model.
}
\examples{
my_rf_cv(5)

}
\keyword{inference}
\keyword{prediction}
