% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test function}
\usage{
my_t.test(x, alternative = "two.sided", mu)
}
\arguments{
\item{x}{Numeric vector input of data.}

\item{alternative}{A character string input that specifies the alternative
hypothesis, defaults to \code{"two.sided"}. Also accepts \code{"less"} and
\code{"greater"}.}

\item{mu}{Numeric input specifying the null hypothesis of the mean.}
}
\value{
List with the following elements: Test statistic, degrees of freedom,
alternative hypothesis, and P-value.
}
\description{
This function performs a t-test on input data and null hypothesis \code{mu}.
}
\examples{
set.seed(111)
x <- rnorm(10, mean = 0, sd = 1)
my_t.test(x, mu = 0)
my_t.test(x, alternative = "less", mu = 1)

}
\keyword{inference}
