% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{StatsTools_packages}
\alias{StatsTools_packages}
\title{List all packages imported by StatsTools}
\usage{
StatsTools_packages()
}
\description{
List all packages imported by StatsTools
}
\examples{
StatsTools_packages()
}
